Separate ManifestStats in its own module file
